+++ twitter4j-v2-support/src/main/kotlin/twitter4j/UserMentionEntity2.kt
@@ -1,10 +1,16 @@
 package twitter4j
 
-data class UserMentionEntity2(val start: Int, val end: Int, val username: String) {
+data class UserMentionEntity2(
+    val start: Int,
+    val end: Int,
+    val screenName: String,
+    val id: Long
+) {
 
     constructor(json: JSONObject) : this(
         json.getInt("start"),
         json.getInt("end"),
-        json.getString("username")
+        json.getString("username"),
+        json.optLong("id", -1L)
     )
 }